+++ buildpacks/nodejs-corepack/src/cmd.rs
@@ -1,9 +1,8 @@
-use bullet_stream::state::SubBullet;
-use bullet_stream::{style, Print};
+use bullet_stream::global::print;
+use bullet_stream::style;
 use fun_run::CommandWithName;
 use heroku_nodejs_utils::vrs::{Version, VersionError};
 use libcnb::Env;
-use std::io::Stderr;
 use std::{path::Path, process::Command};
 
 #[derive(Debug)]
@@ -31,8 +30,7 @@ pub(crate) fn corepack_enable(
     package_manager: &str,
     shim_path: &Path,
     env: &Env,
-    mut log: Print<SubBullet<Stderr>>,
-) -> Result<Print<SubBullet<Stderr>>, fun_run::CmdError> {
+) -> Result<(), fun_run::CmdError> {
     let shim_path_string = shim_path.to_string_lossy();
     let mut command = Command::new("corepack");
     command.args([
@@ -43,24 +41,12 @@ pub(crate) fn corepack_enable(
     ]);
     command.envs(env);
 
-    log = log.sub_bullet(format!("Executing {}", style::command(command.name())));
+    print::sub_bullet(format!("Executing {}", style::command(command.name())));
 
-    command.named_output()?;
-
-    Ok(log)
+    command.named_output().map(|_| ())
 }
 
 /// Execute `corepack prepare` to install the correct package manager
-pub(crate) fn corepack_prepare(
-    env: &Env,
-    mut log: Print<SubBullet<Stderr>>,
-) -> Result<Print<SubBullet<Stderr>>, fun_run::CmdError> {
-    let mut command = Command::new("corepack");
-    command.arg("prepare");
-    command.envs(env);
-    log.stream_with(
-        format!("Running {}", style::command(command.name())),
-        |stdout, stderr| command.stream_output(stdout, stderr),
-    )?;
-    Ok(log)
+pub(crate) fn corepack_prepare(env: &Env) -> Result<(), fun_run::CmdError> {
+    print::sub_stream_cmd(Command::new("corepack").arg("prepare").envs(env)).map(|_| ())
 }