+++ buildpacks/nodejs-pnpm-engine/src/main.rs
@@ -5,7 +5,7 @@
 
 mod errors;
 
-use bullet_stream::Print;
+use bullet_stream::global::print;
 use libcnb::build::{BuildContext, BuildResult};
 use libcnb::data::build_plan::BuildPlanBuilder;
 use libcnb::detect::{DetectContext, DetectResult, DetectResultBuilder};
@@ -13,7 +13,6 @@ use libcnb::generic::{GenericMetadata, GenericPlatform};
 use libcnb::{buildpack_main, Buildpack};
 #[cfg(test)]
 use libcnb_test as _;
-use std::io::stderr;
 #[cfg(test)]
 use test_support as _;
 
@@ -40,12 +39,14 @@ impl Buildpack for PnpmEngineBuildpack {
     }
 
     fn build(&self, context: BuildContext<Self>) -> libcnb::Result<BuildResult, Self::Error> {
-        let _logger = Print::new(stderr()).h1(context
-            .buildpack_descriptor
-            .buildpack
-            .name
-            .as_ref()
-            .expect("The buildpack.toml should have a 'name' field set"));
+        print::buildpack(
+            context
+                .buildpack_descriptor
+                .buildpack
+                .name
+                .as_ref()
+                .expect("The buildpack.toml should have a 'name' field set"),
+        );
 
         // This buildpack does not install pnpm yet, suggest using
         // `heroku/nodejs-corepack` instead.