+++ buildpacks/nodejs-yarn/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Buildpack output changes. ([#1093](https://github.com/heroku/buildpacks-nodejs/pull/1093))
+
 ## [3.6.1] - 2025-04-24
 
 ### Added